+++ Employee.cpp
@@ -1,3 +1,3 @@
 #include "Employee.h"
 
-Employee::Employee(int _ID, Company* _company, int _salary, int _grade): ID(_ID), company(_company), salary(_salary), grade(_grade) {}
+Employee::Employee(int _ID, Company* _company, long long int _salary, long long int _grade): ID(_ID), company(_company), salary(_salary), grade(_grade) {}
